refactor(research): split eval scoring into company and market halves
- Grading one company's profile and grading a whole market's list are entirely different code that share only the shapes they are written against, so each now has a file of its own with those shapes beside them.
- Nothing about how a run is scored changed, and the file other modules import still carries the same names, so no caller moved.
