+++ packages/research/src/application/eval-scoring.ts
@@ -28,11 +28,10 @@
  *   duplicate rate       whether the fold that joins two rows of one company still
  *                        holds — see its own note for what it cannot see
  *   location fill        how many rows say what town or province the company is in
- *   confirmation rate    how many rows two independent websites establish as real
  *
- * The row count is still reported, as the scale those five read against rather than
- * as the grade — a pass can hold its confirmation rate up by returning fewer rows,
- * so only the two together say what happened.
+ * The row count is still reported, as the scale those four read against rather than
+ * as the grade — it is also what checking that every row is a real company would
+ * cost, so it needs a reading of its own before that check exists.
  *
  * Grounding is judged by which pages the run *fetched*, not which its findings cite:
  * once per-field citations point at whichever page stated each fact, a run that
@@ -41,811 +40,55 @@
  * A run is also grounded without fetching the target's site at all when an official
  * company-register lookup resolved the target by legal name — an independent proof
  * the right entity was reached.
+ *
+ * The two halves this rolls up live beside it: eval-scoring-company.ts grades one
+ * company's profile, eval-scoring-market.ts grades a market's list, and the shapes both
+ * are written against are in eval-scoring-types.ts.
  */
-
-import { foldLabel } from '@batuda/domain'
 
 import {
-	branchOfficeParents,
-	discoveryRowIdentityKeys,
-} from './prospect-dedupe-guard'
-import { anyTermAppearsIn, foldDiacritics, termTokens } from './term-match'
+	contactNameMatches,
+	fieldMatches,
+	isFilled,
+	normalizeDomain,
+	specificLocationAgrees,
+} from './eval-scoring-company'
+import {
+	duplicatedRows,
+	isKnownNonCompany,
+	partsAnsweredBy,
+} from './eval-scoring-market'
+import {
+	type EvalSummary,
+	endedWithAnAnswer,
+	type GoldenExpectation,
+	isSucceeded,
+	type MarketScore,
+	type RunOutcome,
+	type RunScore,
+	SCORABLE_FIELDS,
+} from './eval-scoring-types'
 
-/**
- * The enrichment scalars we can check against an objective golden answer. Free-text
- * fields (current tools, tags) are left out because there is no single correct
- * value to score them against. Snake case matches the extraction schema's field
- * names.
- *
- * A company's mailbox, telephone number and registration number are in because each
- * has exactly one right answer that can be written down and checked. Adding them
- * costs nothing in the two per-field figures — a row that states no expected value
- * for a field is skipped for that field, in both the precision and the recall
- * count — but it does change the empty rate, and deliberately so: a run whose only
- * real find was the mailbox printed on the company's contact page used to be filed
- * alongside the runs that found nothing at all.
- *
- * The company's website is deliberately NOT in. It is already the thing grounding is
- * measured on — a golden row's `officialDomain` — so scoring it here would report
- * the same success twice and make a change to grounding look twice as large.
- */
-export const SCORABLE_FIELDS = [
-	'industry',
-	'size_range',
-	'country',
-	'location',
-	'email',
-	'phone',
-	'tax_id',
-] as const
-
-export type ScorableField = (typeof SCORABLE_FIELDS)[number]
-
-/**
- * Size/reach buckets a golden company can be tagged with, so a regression that
- * hits only one segment (a niche or small company, the hardest to research) shows
- * up in a per-bucket breakdown instead of being averaged away. `big` = a
- * household name; `small` = an SMB with a light web presence; `niche` = a
- * specialist with little third-party coverage.
- */
-export const GOLDEN_BUCKETS = ['big', 'small', 'niche'] as const
-
-export type GoldenBucket = (typeof GOLDEN_BUCKETS)[number]
-
-/** The run outcomes that end a run; a run in any of these is what the eval scores. */
-export type TerminalStatus =
-	| 'succeeded'
-	| 'succeeded_low_confidence'
-	| 'no_reliable_data'
-	| 'failed'
-	| 'cancelled'
-
-/** Succeeded-class outcomes — a thin (low-confidence) success still carries findings to score. */
-export const isSucceeded = (status: TerminalStatus): boolean =>
-	status === 'succeeded' || status === 'succeeded_low_confidence'
-
-/**
- * Whether the run got far enough to say something about what it was asked.
- *
- * Wider than succeeding. A search that looked and found nothing ends as no
- * reliable data, and that is an answer about the market — the one a change that
- * empties a search has to be caught by. A run that died or was stopped never
- * reached an answer, and reading its empty hands as "this market has nothing in
- * it" would blame the market for an outage.
- */
-const endedWithAnAnswer = (status: TerminalStatus): boolean =>
-	status !== 'failed' && status !== 'cancelled'
-
-/**
- * One part of a request that asks for several — a trade, a service, a segment —
- * with the wordings that place a returned row in it.
- *
- * The wordings are golden data, never a list in here. Which trades a market has is
- * a fact about the world, so no list of the trades or countries expected of a run
- * belongs in code. A golden set is where the expected answers already live, so they
- * cost nothing there.
- */
-export interface MarketPart {
-	readonly id: string
-	/** Wordings that place a row in this part, in whichever languages the market answers in. */
-	readonly terms: ReadonlyArray<string>
-}
-
-/**
- * What a request for a whole market asks for, when a golden row is a market rather
- * than one company. A scan answers with a list, so nothing a profile is graded on —
- * reaching one company's site, filling that company's fields — is a question here.
- */
-export interface MarketExpectation {
-	/** What this market is called, and the key a per-market breakdown groups on. */
-	readonly name: string
-	/** The parts the request names; coverage is how many came back with a row. */
-	readonly parts: ReadonlyArray<MarketPart>
-	/**
-	 * Organisations known not to be companies in this market — the trade bodies,
-	 * federations and system operators a search for a trade runs straight through
-	 * on its way to the members.
-	 *
-	 * Naming them is what makes the count evidence rather than an opinion, and it is
-	 * the only thing that can see past the organisation-kind guard's own languages:
-	 * that guard reads Spanish, Catalan and English, so asking it again here would
-	 * agree with itself and report every market clean. A body this list does not name
-	 * still passes unmeasured, which is the honest limit of the figure.
-	 */
-	readonly notCompanies: ReadonlyArray<string>
-}
-
-/**
- * One company's known-correct answer. This lives in code as a type; the actual
- * rows live in the eval dataset (Latitude), so a new company is added there, not
- * here.
- */
-export interface GoldenExpectation {
-	readonly id: string
-	/** What the run is asked to research — a name, or name + location for a generic name. */
-	readonly query: string
-	/** The company's own official website host — the primary proof the target was reached. */
-	/**
-	 * The company's own website host — the primary proof the run reached the target.
-	 * Null for a company that has no site of its own, where the proof is an alt
-	 * domain (a register entry, a directory page) or a register lookup instead.
-	 */
-	readonly officialDomain: string | null
-	/** Other hosts that still prove the target was reached (a registry profile, a known subsidiary). */
-	readonly altDomains?: ReadonlyArray<string>
-	/** Known-correct field values. Only fields listed here are scored for precision. */
-	readonly fields: Partial<Record<ScorableField, string>>
-	/** People the company is known to publish; scores whether the run recovered them
-	 * *with a title* — the recall the focused contacts pass exists to lift. */
-	readonly contacts?: ReadonlyArray<{ readonly name: string }>
-	/** Size/reach segment, so quality can be reported per bucket, not just overall. */
-	readonly bucket?: GoldenBucket
-	/**
-	 * Present when this row asks for a market rather than one company. The two are
-	 * one type because a pass runs one or the other and every figure below already
-	 * reports nothing when it had nothing to judge.
-	 */
-	readonly market?: MarketExpectation
-}
-
-/**
- * What a run came back with, counted rather than judged against a known answer.
- * Without this, a run that returns four right scalars and nothing else scores
- * the same as one that returns a full picture — the opposite of what a rich
- * profile is worth.
- */
-export interface ProfileFullness {
-	/** Profile fields the shape asks for, and how many carry a real value. */
-	readonly fieldsTotal: number
-	readonly fieldsFilled: number
-	/** People named, whether or not a title came with them. */
-	readonly contactsNamed: number
-	/** Of those, how many carry a title — the ones worth writing to. */
-	readonly contactsTitled: number
-}
-
-/**
- * A normalized view of what one run produced, adapted from the run row + its
- * findings by the caller. Keeping the adapter out of here is deliberate: the
- * findings shape changes when citations move per-field, but this view — and so
- * every metric below — does not.
- */
-export interface RunOutcome {
-	readonly status: TerminalStatus
-	/** What the run was billed and what it consumed; absent when not read back. */
-	readonly usage?: RunUsage
-	/**
-	 * Hosts of the sources the run fetched — the pages it actually reached. The run
-	 * reaching the target's own site is what proves it researched the right company;
-	 * once that is confirmed, per-field citations are free to point at whichever
-	 * fetched page stated each fact (a registry, a directory), not only the homepage.
-	 */
-	readonly reachedDomains: ReadonlyArray<string>
-	/** Extracted enrichment scalars; a missing/blank value counts as unfilled. */
-	readonly fields: Partial<Record<ScorableField, string | null>>
-	/** Contacts the run extracted (after the guards), each with the title it found
-	 * or null — a named person recovered without a title still counts as titleless. */
-	readonly contacts: ReadonlyArray<{
-		readonly name: string
-		readonly role: string | null
-	}>
-	/**
-	 * The companies a discovery scan came back with. Empty for a run that answers
-	 * with a profile rather than a list. This is what a scan is asked for, so it is
-	 * what a scan has to be scored on — each row carrying the four things the market
-	 * figures read: who it is, where it is, what it says it does, and its web
-	 * address, which is one of the two things that tell two rows apart.
-	 */
-	readonly companies: ReadonlyArray<{
-		readonly name: string
-		/** The address the row carries, as written, or null when it carries none. */
-		readonly website: string | null
-		/** The town or province the row states, or null when it states none. */
-		readonly location: string | null
-		/** The row's own words about what it is and why it matched, run together. */
-		readonly describedAs: string
-		/**
-		 * Whether two independent websites established the company, one of them its
-		 * own site. False on a row carrying no verdict, which is the right reading
-		 * rather than a missing one: nothing established that company.
-		 */
-		readonly confirmed: boolean
-	}>
-	/**
-	 * Whether an official-registry lookup this run resolved the target company by
-	 * its legal name. Independent of the fetched pages: a company confirmed in the
-	 * register was reached even if its own site was never scraped.
-	 */
-	readonly registryConfirmed?: boolean
-	/** How full the profile came back, across the whole output shape. */
-	readonly profile?: ProfileFullness
-}
-
-// What one run was billed, and what it consumed getting there — read back from
-// the run's own row, so it is what was recorded rather than a live tally.
-export interface RunUsage {
-	/** Everything the run was billed, in whole cents. */
-	readonly costCents: number
-	/** Of that, what the metered per-call lookups cost. */
-	readonly paidCostCents: number
-	readonly tokensIn: number
-	readonly tokensOut: number
-	/** Provider credits consumed, totalled across every provider. */
-	readonly creditsUsed: number
-	/**
-	 * The models that answered this run, keyed `<tier>@<model>` with the number
-	 * of calls each took. A tier listed twice fell back partway through — which
-	 * matters when reading the score, since the run was then partly the work of
-	 * a model the pass was not set up to measure.
-	 */
-	readonly callsByModel?: Record<string, number>
-}
-
-/**
- * What one scan's list got right, counted against the market it was asked for.
- *
- * Counts rather than rates, so a pass adds them up and then divides once: a market
- * that came back with sixty rows then weighs sixty rows against a six-row market's
- * six, instead of each contributing one equal rate to an average.
- */
-export interface MarketScore {
-	/** Which market this run answered for — the key a per-market breakdown groups on. */
-	readonly name: string
-	/**
-	 * Rows the scan came back with. The scale the figures below read against, and
-	 * what checking that every row is a real company would cost, so it is reported
-	 * in its own right rather than only as a denominator.
-	 */
-	readonly rowsReturned: number
-	/** Of those, how many are not an organisation the golden names as not a company. */
-	readonly rowsRightKind: number
-	/**
-	 * Of the rows returned, how many the run established as real companies.
-	 *
-	 * Counted over every row rather than the right-kind ones, so it answers "how
-	 * much of what came back does the run stand behind" — the figure the recall
-	 * cost of requiring two independent sources is read off. A pass whose rows
-	 * carry no verdict reads nought, which is what it confirmed.
-	 */
-	readonly rowsConfirmed: number
-	/**
-	 * Of the rows returned — every one of them, including any the golden names as not
-	 * a company — how many say where the company is. The field is asked for a town or
-	 * a province, but any stated place counts here — a row answering with the country
-	 * alone is filled as far as this can tell, and the request having named that
-	 * country is not something a count of rows can know.
-	 */
-	readonly rowsLocated: number
-	/** How many rows are another row's company written a second time. */
-	readonly rowsDuplicated: number
-	/** Parts the request named — the trades, services or segments it asked for. */
-	readonly partsExpected: number
-	/** Of those, how many came back with at least one row. */
-	readonly partsAnswered: number
-}
-
-export interface RunScore {
-	readonly id: string
-	/** What this run cost, carried through so a pass can be totalled and compared. */
-	readonly usage?: RunUsage
-	readonly grounded: boolean
-	/**
-	 * Whether reaching a particular company was a question for this row at all. A
-	 * market names no company to reach, so counting it as a grounding failure would
-	 * report a whole scan pass at zero for answering the question it was asked.
-	 */
-	readonly groundable: boolean
-	readonly wrongCompany: boolean
-	/**
-	 * Whether a wrong company got as far as finishing clean — the most that could
-	 * ever reach a record unwatched. An upper bound, not a count of what would:
-	 * writing anything without a person also needs the organisation to have asked
-	 * for it, the value to be a way of reaching somebody rather than a judgement,
-	 * and that address to have come back reachable. None of which is knowable
-	 * here. Read it as "how much even had the chance", and expect the real number
-	 * to be far lower.
-	 */
-	readonly wrongCompanyAutoApplicable: boolean
-	/** Whether the run finished flagged as needing somebody to read it. */
-	readonly lowConfidence: boolean
-	readonly empty: boolean
-	/** Golden fields we have a true answer for (recall's denominator). */
-	readonly fieldsExpected: number
-	/** Golden fields the run actually filled, so we could check them (precision's denominator). */
-	readonly fieldsScored: number
-	/** Of the filled-and-checkable fields, how many matched (the shared numerator). */
-	readonly fieldsCorrect: number
-	/** Known-published people we expected the run to recover (contact recall's denominator). */
-	readonly contactsExpected: number
-	/** Of those, how many the run returned *with a title* (contact recall's numerator). */
-	readonly contactsFound: number
-	/** The golden row's size/reach bucket, carried through so scores can be grouped. */
-	readonly bucket?: GoldenBucket
-	/** The golden row's expected country (ISO alpha-2), for a by-country breakdown. */
-	readonly country?: string
-	/** How full the profile came back, independent of the golden answers. */
-	readonly profile?: ProfileFullness
-	/** What the list got right, present only for a row that asked for a market. */
-	readonly market?: MarketScore
-}
-
-export interface EvalSummary {
-	readonly runs: number
-	/**
-	 * Of the runs that were asked to reach a particular company, the share that
-	 * reached it. Null for a pass of market requests, which name no company to reach
-	 * — the alternative reads zero and says the pass failed at something nobody asked
-	 * it to do.
-	 */
-	readonly groundingAccuracy: number | null
-	/**
-	 * The look-alike figures, over the same runs as grounding and null for the same
-	 * reason: a market request answers with a list, so there is no other company it
-	 * could have come back with instead.
-	 */
-	readonly wrongCompanyRate: number | null
-	/**
-	 * Share of runs that shipped another company's data and finished clean — the
-	 * most that could ever be written with nobody looking. An upper bound: see
-	 * `wrongCompanyAutoApplicable` for the conditions it cannot see.
-	 */
-	readonly wrongCompanyAutoApplicableRate: number | null
-	/** Share of runs that finished flagged as needing somebody to read them. */
-	readonly lowConfidenceRate: number
-	readonly emptyRate: number
-	/** Micro-averaged correct ÷ filled across all runs; null when nothing was filled to check. */
-	readonly fieldPrecision: number | null
-	/** Micro-averaged correct ÷ known across all runs; null when the golden set specified no fields. */
-	readonly fieldRecall: number | null
-	/** Micro-averaged known people recovered *with a title* ÷ known people, across all
-	 * runs; null when no golden row listed expected contacts. */
-	readonly contactRecall: number | null
-	/**
-	 * The four market figures, each micro-averaged over every market request in the
-	 * pass rather than averaged across markets, so a sixty-row market weighs sixty
-	 * rows against a six-row one's six.
-	 *
-	 * Null means there was nothing to divide by, which happens two ways, and the row
-	 * count below tells them apart: no market request in the pass at all, where the
-	 * row count is null too, or markets that ran and came back with no rows, where it
-	 * is nought. Coverage is the exception — it divides by the parts a request named,
-	 * so a market that found nothing still reads nought rather than nothing.
-	 */
-	readonly organisationKindPrecision: number | null
-	readonly requestCoverage: number | null
-	readonly duplicateRate: number | null
-	readonly locationFill: number | null
-	/**
-	 * Of every row that came back, the share the run established as a real
-	 * company. This is where requiring two independent sources shows up: it is
-	 * meant to fall below the row count, and how far is the cost of the check.
-	 * Read it beside `rowsPerScan` — a pass can hold its confirmation rate up by
-	 * returning fewer rows, and only the two together say what happened.
-	 */
-	readonly confirmationRate: number | null
-	/**
-	 * Rows one market request came back with on average. Not a quality figure —
-	 * it is the scale the four above read against, and the reading that says what
-	 * a change bought or cost in results.
-	 */
-	readonly rowsPerScan: number | null
-	/**
-	 * Profile fields filled per run, averaged over the runs that reported it, and
-	 * the shape's own field count to read it against.
-	 */
-	readonly fieldsFilledPerRun: number | null
-	readonly profileFieldsTotal: number | null
-	/** People named per run, and how many of those carry a title. */
-	readonly contactsNamedPerRun: number | null
-	readonly contactsTitledPerRun: number | null
-	/** What one run cost on average, in cents; null when no run reported a cost. */
-	readonly costPerRun: number | null
-	/**
-	 * What one usable run cost: what the runs asked to reach a company spent, over
-	 * how many of them reached it, so the ones that found nothing count as waste
-	 * rather than being ignored. A market request is outside both halves — it is not
-	 * asked to reach anybody, and its spend billed to the company runs would read as
-	 * a cost blow-up rather than as the different job it is.
-	 */
-	readonly costPerGroundedRun: number | null
-	/** Of the average run cost, what the metered per-call lookups accounted for. */
-	readonly paidCostPerRun: number | null
-	readonly tokensPerRun: number | null
-	/** Provider credits one run consumed on average. */
-	readonly creditsPerRun: number | null
-	/**
-	 * How many calls each model answered across the whole pass, keyed
-	 * `<tier>@<model>`. Read this before the quality numbers: a tier that shows
-	 * up under two models did not measure the one it was set up to measure, and
-	 * the split says how far off that reading is.
-	 */
-	readonly callsByModel: Record<string, number>
-	/**
-	 * The share of runs where some tier answered on more than one model. Zero
-	 * means every run stayed on its first choice and the scores speak for it.
-	 */
-	readonly cascadedRunRate: number | null
-}
-
-/**
- * Bare registrable host, lowercased, without scheme / `www.` / path — so
- * "https://www.Acme.com/contact" and "acme.com" compare equal.
- */
-const normalizeDomain = (value: string): string => {
-	const withoutScheme = value
-		.trim()
-		.toLowerCase()
-		.replace(/^[a-z][a-z0-9+.-]*:\/\//, '')
-	const host = withoutScheme.split('/')[0] ?? withoutScheme
-	return host.replace(/^www\./, '')
-}
-
-/** Case- and space-insensitive text, for comparing enum-ish field values. */
-export const normalizeText = (value: string): string =>
-	value.trim().toLowerCase().replace(/\s+/g, ' ')
-
-const isFilled = (value: string | null | undefined): value is string =>
-	typeof value === 'string' && value.trim().length > 0
-
-// Titles a page prints before a name ("Sir James Dyson", "Dr Jane Roe"): not part of
-// the name, so a leading one is dropped before matching. "Don"/"Doña" are deliberately
-// absent — "Don" is also a real given name (Don Draper), and dropping it would lose a
-// real person. Only stripped when two tokens remain, so a short "Dr Dre" keeps both.
-const HONORIFICS = new Set([
-	'sir',
-	'dame',
-	'lord',
-	'lady',
-	'mr',
-	'mrs',
-	'ms',
-	'miss',
-	'mx',
-	'dr',
-	'prof',
-	'professor',
-	'rev',
-	'reverend',
-	'hon',
-	'madam',
-	'madame',
-])
-
-// Everyday short forms folded to the formal name a company usually publishes, so a
-// golden "Pete Roever" matches a run's "Peter Roever". Kept small and English; since a
-// match still needs the surname too (≥2 tokens), folding a first name can't collapse
-// two genuinely different people.
-const NICKNAMES: Record<string, string> = {
-	pete: 'peter',
-	rob: 'robert',
-	bob: 'robert',
-	robbie: 'robert',
-	bill: 'william',
-	billy: 'william',
-	tony: 'anthony',
-	jim: 'james',
-	jimmy: 'james',
-	mike: 'michael',
-	dave: 'david',
-	tom: 'thomas',
-	tommy: 'thomas',
-	dan: 'daniel',
-	danny: 'daniel',
-	dick: 'richard',
-	rick: 'richard',
-	matt: 'matthew',
-	greg: 'gregory',
-	ben: 'benjamin',
-	ed: 'edward',
-	eddie: 'edward',
-	andy: 'andrew',
-	ron: 'ronald',
-	ken: 'kenneth',
-	joe: 'joseph',
-	steve: 'steven',
-	nick: 'nicholas',
-	tim: 'timothy',
-	charlie: 'charles',
-}
-
-const stripHonorific = (
-	tokens: ReadonlyArray<string>,
-): ReadonlyArray<string> =>
-	tokens.length > 2 && HONORIFICS.has(tokens[0] ?? '')
-		? tokens.slice(1)
-		: tokens
-
-// A person's name as its accent-folded, lower-cased word tokens, with a leading
-// honorific dropped and each token folded to its formal form. Single-character tokens
-// (a middle initial) are kept, so two people who differ only by initial stay distinct.
-const nameTokens = (name: string): ReadonlyArray<string> => {
-	const rawTokens = foldDiacritics(name)
-		.toLowerCase()
-		.split(/[^a-z0-9]+/)
-		.filter(token => token.length > 0)
-	return stripHonorific(rawTokens).map(token => NICKNAMES[token] ?? token)
-}
-
-// Two names refer to the same person when the shorter one's tokens are all in the
-// longer's — so "Andrew Smith" matches "Andrew J. Smith" without matching a
-// different Smith. A lone shared token (just a first name) is too weak to confirm
-// the same person, so the shorter name must carry at least two tokens. Shared with
-// the contact-discovery eval, so both agree on what "same person" means. Conservative
-// on purpose: an unmatched real person is a miss the eval should show, not paper over.
-export const contactNameMatches = (
-	expected: string,
-	actual: string,
-): boolean => {
-	const e = nameTokens(expected)
-	const a = nameTokens(actual)
-	const [small, big] = e.length <= a.length ? [e, new Set(a)] : [a, new Set(e)]
-	if (small.length < 2) return false
-	return small.every(token => big.has(token))
-}
-
-/**
- * A trade is whatever the page calls it, on both sides: the golden holds the
- * wording a person would expect, and the pipeline keeps the wording it read. So
- * this asks whether the two namings are the same trade, not whether they are the
- * same string.
- *
- * `foldLabel` is the same rule the CRM uses to decide that two spellings are one
- * trade, so a difference this accepts is a difference that would land on one
- * entry there. On top of it, a shared stem — a prefix at least half the expected
- * name's length, minimum four characters, starting some word of what was read —
- * covers the endings a language puts on the same word ("fusteria" against
- * "fusteries", "manufacturing" against "manufacturer").
- *
- * A real miss still reads as one: a bank reported as "banking" against an
- * expected "insurance" shares no stem, which is the signal the eval is for.
- */
-const industryMatches = (expected: string, actual: string): boolean => {
-	const foldedExpected = foldLabel(expected)
-	const foldedActual = foldLabel(actual)
-	if (foldedExpected.length === 0 || foldedActual.length === 0) return false
-	if (
-		foldedActual.includes(foldedExpected) ||
-		foldedExpected.includes(foldedActual)
-	)
-		return true
-	const stem = foldedExpected.slice(
-		0,
-		Math.max(4, Math.ceil(foldedExpected.length / 2)),
-	)
-	return foldedActual.split(' ').some(word => word.startsWith(stem))
-}
-
-const fieldMatches = (
-	field: ScorableField,
-	expected: string,
-	actual: string,
-): boolean => {
-	const normalizedExpected = normalizeText(expected)
-	const normalizedActual = normalizeText(actual)
-	if (normalizedExpected.length === 0 || normalizedActual.length === 0)
-		return false
-	// A location is written differently by every source (order, abbreviations,
-	// postcode placement), so accept either string containing the other rather than
-	// demanding a character-exact match.
-	if (field === 'location') {
-		return (
-			normalizedActual.includes(normalizedExpected) ||
-			normalizedExpected.includes(normalizedActual)
-		)
-	}
-	if (field === 'industry') {
-		return industryMatches(normalizedExpected, normalizedActual)
-	}
-	// A telephone number is printed a dozen ways for the same line — spaces, dots,
-	// brackets, a country code on one page and not the next — so only the digits
-	// are compared, and only the last nine of them, which is a full national number
-	// everywhere the pipeline researches. Anything shorter is compared whole.
-	if (field === 'phone') {
-		const lastDigits = (value: string): string => {
-			const digits = value.replace(/\D/g, '')
-			return digits.length > 9 ? digits.slice(-9) : digits
-		}
-		const expectedDigits = lastDigits(expected)
-		return expectedDigits !== '' && lastDigits(actual) === expectedDigits
-	}
-	// A registration number is written with and without its punctuation and its
-	// country prefix, so it is compared on its letters and digits alone.
-	if (field === 'tax_id') {
-		const bare = (value: string): string =>
-			value.replace(/[^a-z0-9]/gi, '').toUpperCase()
-		const expectedBare = bare(expected)
-		return expectedBare !== '' && bare(actual) === expectedBare
-	}
-	// An email address, a country code and a size band are all values the pipeline
-	// is meant to emit verbatim, so they hold to an exact match.
-	return normalizedActual === normalizedExpected
-}
-
-// Global cities so common that a company being there says little about which company
-// it is — a location match on one of these is too weak to confirm the run reached the
-// right company. Normalized once so lookups match the folded field value.
-const MEGACITIES = new Set(
-	[
-		'london',
-		'paris',
-		'madrid',
-		'barcelona',
-		'berlin',
-		'milan',
-		'rome',
-		'new york',
-		'new york city',
-		'nyc',
-		'los angeles',
-		'chicago',
-		'san francisco',
-		'boston',
-		'dublin',
-		'amsterdam',
-		'lisbon',
-		'tokyo',
-		'singapore',
-		'hong kong',
-		'shanghai',
-		'beijing',
-		'mumbai',
-		'delhi',
-		'mexico city',
-		'sao paulo',
-		'buenos aires',
-		'toronto',
-		'sydney',
-	].map(normalizeText),
-)
-
-// Whether the run's location matches the golden's AND that location is specific enough
-// to identify the company (not a global capital). Used only to judge that a run
-// reached the right company when it never touched the official domain — a stronger
-// signal than a coarse industry code, which is shared by many firms.
-const specificLocationAgrees = (
-	expected: GoldenExpectation,
-	outcome: RunOutcome,
-): boolean => {
-	const goldenLocation = expected.fields.location
-	const actual = outcome.fields.location
-	if (goldenLocation === undefined || !isFilled(actual)) return false
-	// Judge the city (the part before any region/country) against the megacity list,
-	// so "London, UK" is recognised as generic just as bare "London" is.
-	const goldenCity = normalizeText(
-		goldenLocation.split(',')[0] ?? goldenLocation,
-	)
-	if (MEGACITIES.has(goldenCity)) return false
-	return fieldMatches('location', goldenLocation, actual)
-}
-
-/**
- * How many of a request's parts came back with at least one row that answers them.
- *
- * Only rows that are the kind of organisation asked for are read. A trade's own
- * federation names that trade in its title, so counting it would mark the part
- * answered by the very body that has no work to sell — and a list holding the solar
- * association and no solar company is the answer this figure exists to catch.
- */
-const partsAnsweredBy = (
-	rows: RunOutcome['companies'],
-	parts: ReadonlyArray<MarketPart>,
-): number => {
-	const rowWords = rows.map(row => termTokens(`${row.name} ${row.describedAs}`))
-	return parts.filter(part => anyTermAppearsIn(part.terms, rowWords)).length
-}
-
-/**
- * Whether a row is one of the organisations the golden names as not a company.
- *
- * The listed name has to appear in the row's name as those words, in that order,
- * next to each other. A run writes a body's name longer than the golden does —
- * "FENIE — Federación Nacional de Empresarios de Instalaciones" — so the listed name
- * sitting inside the row's is what catches it.
- *
- * Whole words rather than a run of letters, because a body is usually known by its
- * initials and three letters land inside an unrelated name by accident: "RTE" sits
- * in the middle of "Norte Instalaciones", and "FFB" inside "Groupe FFBat".
- *
- * Next to each other rather than merely present, because the words of a body's name
- * are the trade's ordinary words scattered through many a company's: an unordered
- * test reads "Eléctrica del Norte, Red de Instaladores" as the grid operator. Order
- * does not save every case — a listed name short enough to sit inside a real company
- * name still matches it — which is why the golden lists the name a body is known by
- * rather than a fragment of it. And only in that direction — asking whether the row's words all
- * sit inside the listed name marks any company named after its own trade as a body,
- * from "Instalaciones y Energía" to a French "Génie Électrique et Climatique".
- * Counting a body where there is none marks a real company as the wrong kind, which
- * overstates the very problem being measured.
- *
- * Two rules on the golden data follow. List the name a body is actually known by,
- * specific enough to be its own — "Red Eléctrica" alone names half the installers in
- * the country — and list its initials as an entry of their own, because an acronym
- * shares no words with what it stands for.
- *
- * What is left over, and cannot be read off a name: a company trading under a body's
- * initials, like the retailer FENIE Energía beside the federation FENIE. It reads as
- * the body. Asking the model what an organisation is would settle it; a name cannot.
- */
-const isKnownNonCompany = (
-	name: string,
-	notCompanies: ReadonlyArray<string>,
-): boolean => {
-	const words = termTokens(name)
-	return notCompanies.some(listed => {
-		const listedWords = termTokens(listed)
-		if (listedWords.length === 0) return false
-		// A body known by its initials shares that word with the companies trading
-		// under it — the retailer FENIE Energía beside the federation FENIE, Grupo
-		// Unef Solar beside UNEF, RTE Ascenseurs beside the grid operator. One word
-		// on its own is only conclusive when it is the whole of what the row is
-		// called; spelled-out names carry enough words to be found inside a longer
-		// one safely.
-		if (listedWords.length === 1)
-			return words.length === 1 && words[0] === listedWords[0]
-		return words.some((_, at) =>
-			listedWords.every((word, offset) => words[at + offset] === word),
-		)
-	})
-}
-
-/**
- * How many rows are another row's company a second time: the rows, less the
- * companies they turn out to be.
- *
- * Two rows are one company when they share a name with the legal form off the end,
- * share a site host, or one reads as the other's branch office — the same three ways
- * the scan itself folds rows, and sameness carries across all of them.
- *
- * Reusing what the fold uses is what this number is worth and what bounds it. The
- * list it reads has already been folded that way, so nothing it can see should be
- * left: it answers "is that fold still running and still doing its job", and it moves
- * the moment the answer is no. It cannot answer whether those three ways are the
- * right ones — a repeat of a shape none of them describes reads here as two
- * companies, and counting that needs a hand-marked answer to score against, which is
- * a different measurement from this one.
- *
- * A row nothing can be read from — a name that is all punctuation, no address —
- * files under no key and so counts as its own company. That is the safe direction:
- * it never claims a duplicate it cannot show.
- */
-const duplicatedRows = (rows: RunOutcome['companies']): number => {
-	// The fields the fold reads a row by, in the shape it reads them: the name, the
-	// site, and the place — a branch is told by its name ending on the town it gives.
-	const asDiscoveryRows = rows.map(row => ({
-		name: row.name,
-		...(row.website === null ? {} : { website: row.website }),
-		...(row.location === null ? {} : { location: row.location }),
-	}))
-	const parentOfBranch = branchOfficeParents(asDiscoveryRows)
-
-	// Each company found so far, as the keys its rows filed under. A row meeting one
-	// of them is that company again; a row meeting two proves those two were one
-	// company all along, so they merge. A branch is filed under the company it hangs
-	// off as well as itself, which is how it meets it whichever of the two came first.
-	const companies: Array<Set<string>> = []
-	for (const [at, row] of asDiscoveryRows.entries()) {
-		const parent = parentOfBranch.get(at)
-		const parentRow = parent === undefined ? undefined : asDiscoveryRows[parent]
-		const keys = [
-			...discoveryRowIdentityKeys(row),
-			...(parentRow === undefined ? [] : discoveryRowIdentityKeys(parentRow)),
-		]
-		const matched = companies.filter(company =>
-			keys.some(key => company.has(key)),
-		)
-		const mergeInto = matched[0]
-		if (mergeInto === undefined) {
-			companies.push(new Set(keys))
-			continue
-		}
-		for (const key of keys) mergeInto.add(key)
-		for (const other of matched.slice(1)) {
-			for (const key of other) mergeInto.add(key)
-			companies.splice(companies.indexOf(other), 1)
-		}
-	}
-	return rows.length - companies.length
-}
+export { contactNameMatches } from './eval-scoring-company'
+export {
+	type EvalSummary,
+	GOLDEN_BUCKETS,
+	type GoldenBucket,
+	type GoldenExpectation,
+	isSucceeded,
+	type MarketExpectation,
+	type MarketPart,
+	type MarketScore,
+	normalizeText,
+	type ProfileFullness,
+	type RunOutcome,
+	type RunScore,
+	type RunUsage,
+	SCORABLE_FIELDS,
+	type ScorableField,
+	type TerminalStatus,
+} from './eval-scoring-types'
+export { foldDiacritics, termTokens } from './term-match'
 
 /** Score one run against its golden expectation. */
 export const scoreRun = (
@@ -966,8 +209,8 @@ export const scoreRun = (
 			: {
 					name: expectedMarket.name,
 					rowsReturned: outcome.companies.length,
-					rowsRightKind: rightKindRows.length,
 					rowsConfirmed: outcome.companies.filter(row => row.confirmed).length,
+					rowsRightKind: rightKindRows.length,
 					rowsLocated: outcome.companies.filter(row => isFilled(row.location))
 						.length,
 					rowsDuplicated: duplicatedRows(outcome.companies),
@@ -1018,10 +261,10 @@ export const summarizeScores = (
 			fieldRecall: null,
 			contactRecall: null,
 			organisationKindPrecision: null,
+			confirmationRate: null,
 			requestCoverage: null,
 			duplicateRate: null,
 			locationFill: null,
-			confirmationRate: null,
 			rowsPerScan: null,
 			fieldsFilledPerRun: null,
 			profileFieldsTotal: null,
